// src/generated/sensor/camera_sensor.ts
export enum ImageFormat {
  UNKNOWN = 0,
  RGB = 1,
  BGR = 2,
  RGBA = 3,
  BGRA = 4,
  GRAY = 5,
  UNRECOGNIZED = -1,
}
// ...
export enum ImageCompression {
  NONE = 0,
  JPEG = 1,
  PNG = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface ImageData {
  image: Uint8Array;
  height: number;
  width: number;
  format: ImageFormat;
  compression: ImageCompression;
}

function createBaseImageData(): ImageData {
  return { image: new Uint8Array(0), height: 0, width: 0, format: 0, compression: 0 };
}

export const ImageData: MessageFns<ImageData> = {
  encode(message: ImageData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.image.length !== 0) {
      writer.uint32(10).bytes(message.image);
    }
    if (message.height !== 0) {
      writer.uint32(16).int32(message.height);
    }
    if (message.width !== 0) {
      writer.uint32(24).int32(message.width);
    }
    if (message.format !== 0) {
      writer.uint32(32).int32(message.format);
    }
    if (message.compression !== 0) {
      writer.uint32(40).int32(message.compression);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ImageData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseImageData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.image = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.height = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.width = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.format = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.compression = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ImageData {
    return {
      image: isSet(object.image) ? bytesFromBase64(object.image) : new Uint8Array(0),
      height: isSet(object.height) ? globalThis.Number(object.height) : 0,
      width: isSet(object.width) ? globalThis.Number(object.width) : 0,
      format: isSet(object.format) ? imageFormatFromJSON(object.format) : 0,
      compression: isSet(object.compression) ? imageCompressionFromJSON(object.compression) : 0,
    };
  },

  toJSON(message: ImageData): unknown {
    const obj: any = {};
    if (message.image.length !== 0) {
      obj.image = base64FromBytes(message.image);
    }
    if (message.height !== 0) {
      obj.height = Math.round(message.height);
    }
    if (message.width !== 0) {
      obj.width = Math.round(message.width);
    }
    if (message.format !== 0) {
      obj.format = imageFormatToJSON(message.format);
    }
    if (message.compression !== 0) {
      obj.compression = imageCompressionToJSON(message.compression);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ImageData>, I>>(base?: I): ImageData {
    return ImageData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ImageData>, I>>(object: I): ImageData {
    const message = createBaseImageData();
    message.image = object.image ?? new Uint8Array(0);
    message.height = object.height ?? 0;
    message.width = object.width ?? 0;
    message.format = object.format ?? 0;
    message.compression = object.compression ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}